LAMBDA_LABS/niyon-be-chat/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dotenv==0.14.0
python-engineio==3.13.1
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
